Check function

+++ Models/Salary.cs
@@ -4,9 +4,9 @@
     {
         public string? UserId { get; set; }
         public string? ClassId { get; set; }
-        public decimal? RateOfCharge { get; set; }
+        public int? RateOfCharge { get; set; }
         public DateTime CreatedDate { get; set; } = DateTime.Now;
-        public decimal? Allowance { get; set; }
+        public int? Allowance { get; set; }
         public DateTime? UpdateDate { get; set; }  
         public string? Note { get; set; }
         public virtual User? GetUser {  get; set; }
